:ambulance: web_debranding: fix search_count method
The module inherits `base` to redefine search method for `payment.acquirer` (it
could not be done directly because `payment` module may be not installed).
However, the super may return integer if method is used as `search_count`. So,
attempt to user `filtered` method would raise error.

Fix it by skiping `search_count`  case.